upload buttons in their own activity section
